♻️ refactor: Lotto 객체 domain 패키지로 변경
